Crane needs configuration of SSL certs separate from pulp
On the Katello, pulp is manage_apache => false and uses it's own
CA.  On a capsule, pulp is manage_apache => true and uses our
apache certs.  So, we need to configure crane directly to use
apache certs in all cases.

This fixes BATS.  Eventually we need to figure out how to standardize
the pulp configurations so they're more similar.