fixes #17367 - backwards-compatible rhsm proxy on main vhost
This moves the RHSM proxy to the main Apache vhost, so the URL's
for a Foreman proxy and a Katello both use the same port (443 by
default).  But we can't just break existing capsules where many
clients could be pointed to 8443 still.

In order to be backwards compatible with previous installations which
hosted a reverse proxy on 8443, it allows configuration of multiple
ports for the main Apache vhost to listen on.

New installs would default to 443 for everything.

Users have a couple of options for upgrade scenario:

(1)  The default upgrade would require no changes on their part:

        - Separate reverse proxy vhost goes away
        - Pulp vhost listens on 8443, too (installer migration sets this)
        - Clients need no updates

(2) They can also opt to continue listening on 8443, but update the RPM
to point to 443 so new clients point to 443. They can move existing
clients at their convenience and eventually discontinue 8443.

(3) Or they can just stop listening on 8443 entirely and use 443, and
update all of their clients with a new RPM.