Only set up puppet integration if needed
This still isn't quite perfect, but right now I can't think of a better
way to correctly determine if puppet integration is actually needed.
Ideally we would get rid of the $puppet parameter but then we need to
resort to defined('puppet') and that's compile order dependent.

This does mean at worst we pull in the client and not the server.

We can also change the notify to a before because the integration
is just scripts that read files every run. There is no daemon that
caches it.