Correct chaining on puppet master integration
The certs::puppet class inherits from certs so that require is
redundant.

We must also chain it to foreman::puppetmaster because puppet now
contains the entire installation. However, it's also provides the
User[puppet] which is the owner of the file. This creates a dependency
cycle. By setting up the cert foreman::puppetmaster uses before the
actual class is used, we avoid this problem and guarantee the ENC
integration is still set up before the puppet master starts.