Fixes #24974 - don't set ssl_proxy_machine_cert for rhsm proxy
It seems SSLProxyMachineCertificateFile option supports only pkcs1
format (containing BEGIN RSA PRIVATE KEY), while it's not possible to use
this kind of keys in FIPS mode. The implication of this is we can't use
this directive in FIPS enabled environment.

Additionally, I don't see any reason to use this directive in the first
place so it seems it should be safe to remove.