Refs #30363: Generate new bootstrap RPM if old style is detected
Prior implementation of bootstrap RPM created a tarball with
the source script in it. If an RPM is detected with a tarball
source, assume this is an upgrade and generate a new bootstrap
RPM to be based on the new system from this point forward.
This if condition can then be dropped in the next release.