Set up a pub_dir vhost on non-Pulp
In puppet-katello the /pub alias is set up as part of the foreman vhost
on port 80 via katello::pulp. If Pulp 2 is managed by this module, the
/pub alias is set up as part of the Pulp 2 HTTP vhost. That means it's
only needed if Pulp 2 is not present at all, the foreman_proxy_content
vhost is needed. It does mean that katello-client-bootstrap needs to be
ensured in the case where Pulp 2 is present.

This also means the additional pulp-vhost80 includes are not needed.