Fixes #32161: Set disablereuse=on and retry=0 on reverse proxy
This change aims to avoid timeouts and delays when client connection
is interrupted reverse proxying back to the Foreman server.
The reverse proxy is meant for communication back to primary Foreman
server where there are not multiple "workers" to receive a request.
This should reduce the number of clients getting 502 proxy errors
and having to be retried manually thus increasing the reliability.