Fixes #32567: Ensure bootstrap RPM symlink exists
Adds a separate existence check for the symlink to ensure it exists
and points at the latest RPM. This ensures that if something removes
the symlink on the system tne next puppet run will put the symlink
back in place. Prior to this the symlink could be removed from the
system and the only way Puppet would put it back was if the RPM itself
got updated.