Dynamically determine if Puppet should be included
This changes the type to Optional[Boolean]. If it's not explicitly set,
some heuristics are used.

It first looks at if the class is already declared, which means it's
safe to include it (essentially a noop). This depends on the compilation
order.

In order to depend less on the compilation order, it looks up the
key `classes` in Hiera. This key is set by Kafo and means it will be
included eventually. That means it's safe to include as well.

Users who use this module directly (without Kafo) and use the Hiera
`classes` key should set the boolean explicitly.