Fixes #34154 - Refresh on Pulp plugin installation
smart_proxy_pulp dynamically retrieves the Pulp content types and
Katello uses this. This means the features need to be refreshed after a
content type is added. The API also changes so that cache needs to be
regenerated.

This does refreshes on the foreman_smartproxy which could be considered
a private API from theforeman/foreman_proxy and collectors would
silently be skipped if anything changes. That's why this must be covered
by spec tests.