Correct the document root in a standalone deployment
The correct document root for Pulp should be set. /var/www may not be
correct. The default docroot on Red Hat is /var/www/html so setting it
to a higher directory may unintentionally leak files. Pulp actually
defines its own to be certain it's empty. In the old reverse proxy setup
it didn't really matter since / is proxied which means nothing is served
anyway.

Fixes: 76e2a6852d1d2ca33935ccf8a6ab69992c32ec1d