Fixes #36854 - Use HTTP/2 when proxying to Foreman
This aims to use mod_proxy_http2, which utilizes HTTP/2 to a backend
server. It depends on a feature released in puppetlabs-apache 10.1.1.

Overall testing shows the connection is much more stable and reliable
than with HTTP/1.1, especially in higher concurrency. It is recommended
to keep this on, but there is an option to disable it if for some reason
it doesn't work. In the future this same option can be used to support
HTTP/3 when Apache implements that.