Fixes #37620 - Use the Pulpcore docroot for reverse proxy
In 804e193b6a180e00ac2828b33a2c11d9a587e5cb /var/www was used as a
docroot, but since / was proxied to Foreman this wasn't an issue. Since
76e2a6852d1d2ca33935ccf8a6ab69992c32ec1d that is no longer the case and
it's now listed.

Pulpcore's docroot is a known safe directory without irrelevant files
that shouldn't be hosted and correct SELinux labels.

It also disables indexes and multiviews on the directory because some
stupid "security" scanners trip on it. This should be safe because the
directory is not used elsewhere. That way we keep listing as an option
for /pub.

Fixes: 76e2a6852d1d2ca33935ccf8a6ab69992c32ec1d